cloud_function/requirements.txt
google-cloud-monitoring
google-cloud-bigquery
google-cloud-storage
datetime